# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an API for Uni Schedule App.",
        "title": "Uni Schedule API",
        "contact": {},
        "version": "0.1"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
...
        "/classes": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create Class",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Class"
                ],
                "summary": "Create Class",
                "operationId": "classes-create",
                "parameters": [
                    {
                        "description": "Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateClassDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.IDResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/classes/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get Class  By ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Class"
                ],
                "summary": "Get Class By ID",
                "operationId": "classes-get-by-id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Class ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Class"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Delete Class",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Class"
                ],
                "summary": "Delete Class",
                "operationId": "classes-delete",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Class ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.IDResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update Class",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Class"
                ],
                "summary": "Update Class",
                "operationId": "classes-update",
                "parameters": [
                    {
                        "description": "Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.UpdateClassDTO"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Class ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.IDResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/schedules": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create a new schedule",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedule"
                ],
                "summary": "Create Schedule",
                "operationId": "schedule-create",
                "parameters": [
                    {
                        "description": "Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateScheduleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.IDResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/schedules/my": {
            "get": {
                "description": "Get Current User Schedules",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedule"
                ],
                "summary": "Get Current User Schedules",
                "operationId": "schedule-get-my",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.GetMySchedulesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/schedules/slug/{slug}": {
            "get": {
                "description": "Get schedule using slug",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedule"
                ],
                "summary": "Get Schedule By Slug",
                "operationId": "schedule-get-slug",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Schedule Slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.ScheduleView"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/schedules/{id}": {
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Delete schedule",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedule"
                ],
                "summary": "Delete Schedule",
                "operationId": "schedule-delete",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Schedule ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.IDResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update schedule",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedule"
                ],
                "summary": "Update Schedule",
                "operationId": "schedule-update",
                "parameters": [
                    {
                        "description": "Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UpdateScheduleRequest"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Schedule ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.IDResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/schedules/{schedule_id}/classes": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get Schedule's Classes",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Class"
                ],
                "summary": "Get Schedule's Classes",
                "operationId": "classes-get-schedule",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Schedule ID",
                        "name": "schedule_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.GetScheduleClassesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "domain.Class": {
            "type": "object",
            "required": [
                "class_type",
                "id",
                "schedule_id",
                "subject_id",
                "teacher_id"
            ],
            "properties": {
                "class_type": {
                    "$ref": "#/definitions/domain.ClassType"
                },
                "id": {
                    "type": "integer"
                },
                "schedule_id": {
                    "type": "integer"
                },
                "subject_id": {
                    "type": "integer"
                },
                "teacher_id": {
                    "type": "integer"
                }
            }
        },
        "domain.ClassType": {
            "type": "string",
            "enum": [
                "lecture",
                "practice",
                "lab",
                "combined"
            ],
            "x-enum-varnames": [
                "Lecture",
                "Practice",
                "Lab",
                "Combined"
            ]
        },
        "domain.ClassView": {
            "type": "object",
            "required": [
                "class_type",
                "id",
                "subject",
                "teacher"
            ],
            "properties": {
                "class_type": {
                    "$ref": "#/definitions/domain.ClassType"
                },
                "id": {
                    "type": "integer"
                },
                "subject": {
                    "$ref": "#/definitions/domain.SubjectView"
                },
                "teacher": {
                    "$ref": "#/definitions/domain.TeacherView"
                }
            }
        },
        "domain.CreateClassDTO": {
            "type": "object",
            "required": [
                "class_type",
                "schedule_id",
                "subject_id",
                "teacher_id"
            ],
            "properties": {
                "class_type": {
                    "$ref": "#/definitions/domain.ClassType"
                },
                "schedule_id": {
                    "type": "integer"
                },
                "subject_id": {
                    "type": "integer"
                },
                "teacher_id": {
                    "type": "integer"
                }
            }
        },
...
        "domain.Day": {
            "type": "string",
            "enum": [
                "monday",
                "tuesday",
                "wednesday",
                "thursday",
                "friday",
                "saturday"
            ],
            "x-enum-varnames": [
                "Monday",
                "Tuesday",
                "Wednesday",
                "Thursday",
                "Friday",
                "Saturday"
            ]
        },
        "domain.Pagination": {
            "type": "object",
            "required": [
                "limit",
                "offest",
                "pages",
                "total"
            ],
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "offest": {
                    "type": "integer"
                },
                "pages": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "domain.Schedule": {
            "type": "object",
            "required": [
                "id",
                "slug",
                "user_id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "domain.ScheduleEntryView": {
            "type": "object",
            "required": [
                "class_number",
                "day",
                "even",
                "id",
                "is_static",
                "odd"
            ],
            "properties": {
                "class_number": {
                    "type": "integer"
                },
                "day": {
                    "$ref": "#/definitions/domain.Day"
                },
                "even": {
                    "$ref": "#/definitions/domain.ClassView"
                },
                "id": {
                    "type": "integer"
                },
                "is_static": {
                    "type": "boolean"
                },
                "odd": {
                    "$ref": "#/definitions/domain.ClassView"
                }
            }
        },
        "domain.ScheduleView": {
            "type": "object",
            "required": [
                "entries",
                "id",
                "slug",
                "user_id"
            ],
            "properties": {
                "entries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.ScheduleEntryView"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "domain.SubjectView": {
            "type": "object",
            "required": [
                "id",
                "name"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "domain.TeacherView": {
            "type": "object",
            "required": [
                "first_name",
                "id",
                "last_name",
                "surname"
            ],
            "properties": {
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_name": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
        "domain.UpdateClassDTO": {
            "type": "object",
            "required": [
                "class_type",
                "subject_id",
                "teacher_id"
            ],
            "properties": {
                "class_type": {
                    "$ref": "#/definitions/domain.ClassType"
                },
                "subject_id": {
                    "type": "integer"
                },
                "teacher_id": {
                    "type": "integer"
                }
            }
        },
...
        "handler.CreateScheduleRequest": {
            "type": "object",
            "required": [
                "slug"
            ],
            "properties": {
                "slug": {
                    "type": "string"
                }
            }
        },
        "handler.ErrorResponse": {
            "type": "object",
            "required": [
                "error"
            ],
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "handler.GetMySchedulesResponse": {
            "type": "object",
            "required": [
                "data",
                "pagination"
            ],
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Schedule"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/domain.Pagination"
                }
            }
        },
        "handler.GetScheduleClassesResponse": {
            "type": "object",
            "required": [
                "data",
                "pagination"
            ],
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.ClassView"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/domain.Pagination"
                }
            }
        },
...
        "handler.IDResponse": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "handler.UpdateScheduleRequest": {
            "type": "object",
            "required": [
                "slug"
            ],
            "properties": {
                "slug": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}